// plugins/roadie-apis-common/src/apis/SecretsSettings/types.ts
/*
 * Copyright 2022 Larder Software Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
export enum SecretStatusType {
  Available = 'Available',
  Updating = 'Updating',
  Not_Set = 'Not Set',
}

export type Secret = {
  name: string;
  value: string;
  description: string;
  status: SecretStatusType;
  helpUrl?: string;
  isCustom?: boolean;
  createdAt?: string;
  lastModified?: string;
};